css/css-color/parsing/color-computed-color-function.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Color Level 4: Computation of colors using color() function syntax</title>
<link rel="help" href="https://drafts.csswg.org/css-color-4/#color-function">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#resolving-color-function-values">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-color-function-values">
<link rel="author" title="Chris Lilley" href="mailto:chris@w3.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="container">
  <div id="target"></div>
</div>
<style>
  #container {
    color: rgb(255, 0, 0);
  }
</style>
<div id="test"></div>
<script>
for (const colorSpace of [ "srgb", "srgb-linear", "a98-rgb", "rec2020", "prophoto-rgb", "display-p3" ]) {
    test_computed_value("color", `color(${colorSpace} 0% 0% 0%)`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 10% 10% 10%)`, `color(${colorSpace} 0.1 0.1 0.1)`);
    test_computed_value("color", `color(${colorSpace} .2 .2 25%)`, `color(${colorSpace} 0.2 0.2 0.25)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 1)`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 0% 0 0 / 0.5)`, `color(${colorSpace} 0 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 20% 0 10/0.5)`, `color(${colorSpace} 0.2 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 20% 0 10/50%)`, `color(${colorSpace} 0.2 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 400% 0 10/50%)`, `color(${colorSpace} 4 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 50% -160 160)`, `color(${colorSpace} 0.5 -160 160)`);
    test_computed_value("color", `color(${colorSpace} 50% -200 200)`, `color(${colorSpace} 0.5 -200 200)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / -10%)`, `color(${colorSpace} 0 0 0 / 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 110%)`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 300%)`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 50% -200)`, `color(${colorSpace} 0.5 -200 0)`);
    test_computed_value("color", `color(${colorSpace} 50%)`, `color(${colorSpace} 0.5 0 0)`);
    test_computed_value("color", `color(${colorSpace})`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 50% -200 / 0.5)`, `color(${colorSpace} 0.5 -200 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 50% / 0.5)`, `color(${colorSpace} 0.5 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} / 0.5)`, `color(${colorSpace} 0 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 200 200 200)`, `color(${colorSpace} 200 200 200)`);
    test_computed_value("color", `color(${colorSpace} 200 200 200 / 200)`, `color(${colorSpace} 200 200 200)`);
    test_computed_value("color", `color(${colorSpace} -200 -200 -200)`, `color(${colorSpace} -200 -200 -200)`);
    test_computed_value("color", `color(${colorSpace} -200 -200 -200 / -200)`, `color(${colorSpace} -200 -200 -200 / 0)`);
    test_computed_value("color", `color(${colorSpace} 200% 200% 200%)`, `color(${colorSpace} 2 2 2)`);
    test_computed_value("color", `color(${colorSpace} 200% 200% 200% / 200%)`, `color(${colorSpace} 2 2 2)`);
    test_computed_value("color", `color(${colorSpace} -200% -200% -200% / -200%)`, `color(${colorSpace} -2 -2 -2 / 0)`);
    test_computed_value("color", `color(${colorSpace} calc(0.5 + 1) calc(0.5 - 1) calc(0.5) / calc(-0.5 + 1))`, `color(${colorSpace} 1.5 -0.5 0.5 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} calc(50% * 3) calc(-150% / 3) calc(50%) / calc(-50% * 3))`, `color(${colorSpace} 1.5 -0.5 0.5 / 0)`);

    test_computed_value("color", `color(${colorSpace} none none none / none)`, `color(${colorSpace} none none none / none)`);
    test_computed_value("color", `color(${colorSpace} none none none)`, `color(${colorSpace} none none none)`);
    test_computed_value("color", `color(${colorSpace} 10% none none / none)`, `color(${colorSpace} 0.1 none none / none)`);
    test_computed_value("color", `color(${colorSpace} none none none / 0.5)`, `color(${colorSpace} none none none / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / none)`, `color(${colorSpace} 0 0 0 / none)`);
}

for (const colorSpace of [ "xyz", "xyz-d50", "xyz-d65" ]) {
    const resultColorSpace = colorSpace == "xyz" ? "xyz-d65" : colorSpace;

    test_computed_value("color", `color(${colorSpace} 0 0 0)`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 1)`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 1 1 1)`, `color(${resultColorSpace} 1 1 1)`);
    test_computed_value("color", `color(${colorSpace} 1 1 1 / 1)`, `color(${resultColorSpace} 1 1 1)`);
    test_computed_value("color", `color(${colorSpace} -1 -1 -1)`, `color(${resultColorSpace} -1 -1 -1)`);
    test_computed_value("color", `color(${colorSpace} 0.1 0.1 0.1)`, `color(${resultColorSpace} 0.1 0.1 0.1)`);
    test_computed_value("color", `color(${colorSpace} 10 10 10)`, `color(${resultColorSpace} 10 10 10)`);
    test_computed_value("color", `color(${colorSpace} .2 .2 .25)`, `color(${resultColorSpace} 0.2 0.2 0.25)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 0.5)`, `color(${resultColorSpace} 0 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} .20 0 10/0.5)`, `color(${resultColorSpace} 0.2 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} .20 0 10/50%)`, `color(${resultColorSpace} 0.2 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / -10%)`, `color(${resultColorSpace} 0 0 0 / 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 110%)`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 300%)`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 1 1)`, `color(${resultColorSpace} 1 1 0)`);
    test_computed_value("color", `color(${colorSpace} 1)`, `color(${resultColorSpace} 1 0 0)`);
    test_computed_value("color", `color(${colorSpace})`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 1 1 / .5)`, `color(${resultColorSpace} 1 1 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 1 / 0.5)`, `color(${resultColorSpace} 1 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} / 50%)`, `color(${resultColorSpace} 0 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} calc(0.5 + 1) calc(0.5 - 1) calc(0.5) / calc(-0.5 + 1))`, `color(${resultColorSpace} 1.5 -0.5 0.5 / 0.5)`);

    test_computed_value("color", `color(${colorSpace} none none none / none)`, `color(${resultColorSpace} none none none / none)`);
    test_computed_value("color", `color(${colorSpace} none none none)`, `color(${resultColorSpace} none none none)`);
    test_computed_value("color", `color(${colorSpace} 0.2 none none / none)`, `color(${resultColorSpace} 0.2 none none / none)`);
    test_computed_value("color", `color(${colorSpace} none none none / 0.5)`, `color(${resultColorSpace} none none none / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / none)`, `color(${resultColorSpace} 0 0 0 / none)`);
}

    const div = document.querySelector("#test");
    function testColorFunction(description, rule, expectedValue) {
        test(function() {
            div.style.color = "black";
            div.style.color = rule;
            assert_equals(getComputedStyle(div).color, expectedValue);
        }, description);
    }

    // Opaque sRGB in color()
    testColorFunction("sRGB all numbers", "color(srgb 1.00 0.50 0.200)", "color(srgb 1 0.5 0.2)");
    testColorFunction("sRGB all percent", "color(srgb 100% 50% 20%)", "color(srgb 1 0.5 0.2)");
    testColorFunction("sRGB mixed number and percent", "color(srgb 100% 0.5 20%)", "color(srgb 1 0.5 0.2)");
    testColorFunction("sRGB mixed number and percent 2", "color(srgb 1.00 50% 0.2)", "color(srgb 1 0.5 0.2)");
    testColorFunction("sRGB all none", "color(srgb none none none)", "color(srgb none none none)");
    testColorFunction("sRGB number and none", "color(srgb 1.00 none 0.2)", "color(srgb 1 none 0.2)");
    testColorFunction("sRGB percent and none", "color(srgb 100% none 20%)", "color(srgb 1 none 0.2)");
    testColorFunction("sRGB number, percent and none", "color(srgb 100% none 0.2)", "color(srgb 1 none 0.2)");

    // non-unity alpha, sRGB in  color()
    testColorFunction("sRGB with alpha, all numbers", "color(srgb 1.00 0.50 0.200 / 0.6)", "color(srgb 1 0.5 0.2 / 0.6)");
    testColorFunction("sRGB with alpha, all percent", "color(srgb 100% 50% 20% / 60%)", "color(srgb 1 0.5 0.2 / 0.6)");
    testColorFunction("sRGB with alpha, mixed number and percent", "color(srgb 100% 0.5 20% / 0.6)", "color(srgb 1 0.5 0.2 / 0.6)");
    testColorFunction("sRGB with alpha, mixed number and percent 2", "color(srgb 1.00 50% 0.2 / 60%)", "color(srgb 1 0.5 0.2 / 0.6)");
    testColorFunction("sRGB with alpha, all none", "color(srgb none none none / none)", "color(srgb none none none / none)");
    testColorFunction("sRGB with alpha, number and none", "color(srgb 1.00 none 0.2 / none)", "color(srgb 1 none 0.2 / none)");
    testColorFunction("sRGB with alpha, percent and none", "color(srgb 100% none 20% / 30%)", "color(srgb 1 none 0.2 / 0.3)");
    testColorFunction("sRGB with alpha, number, percent and none", "color(srgb 100% none 0.2 / 23.7%)", "color(srgb 1 none 0.2 / 0.237)");

    // Opaque linear-light sRGB in color()
    testColorFunction("Linear-light sRGB all numbers", "color(srgb-linear 1.00 0.50 0.200)", "color(srgb-linear 1 0.5 0.2)");
    testColorFunction("Linear-light sRGB all percent", "color(srgb-linear 100% 50% 20%)", "color(srgb-linear 1 0.5 0.2)");
    testColorFunction("Linear-light sRGB mixed number and percent", "color(srgb-linear 100% 0.5 20%)", "color(srgb-linear 1 0.5 0.2)");
    testColorFunction("Linear-light sRGB mixed number and percent 2", "color(srgb-linear 1.00 50% 0.2)", "color(srgb-linear 1 0.5 0.2)");
    testColorFunction("Linear-light sRGB all none", "color(srgb-linear none none none)", "color(srgb-linear none none none)");
    testColorFunction("Linear-light sRGB number and none", "color(srgb-linear 1.00 none 0.2)", "color(srgb-linear 1 none 0.2)");
    testColorFunction("Linear-light sRGB percent and none", "color(srgb-linear 100% none 20%)", "color(srgb-linear 1 none 0.2)");
    testColorFunction("Linear-light sRGB number, percent and none", "color(srgb-linear 100% none 0.2)", "color(srgb-linear 1 none 0.2)");

    // non-unity alpha, linear-light sRGB in  color()
    testColorFunction("Linear-light sRGB with alpha, all numbers", "color(srgb-linear 1.00 0.50 0.200 / 0.6)", "color(srgb-linear 1 0.5 0.2 / 0.6)");
    testColorFunction("Linear-light sRGB with alpha, all percent", "color(srgb-linear 100% 50% 20% / 60%)", "color(srgb-linear 1 0.5 0.2 / 0.6)");
    testColorFunction("Linear-light sRGB with alpha, mixed number and percent", "color(srgb-linear 100% 0.5 20% / 0.6)", "color(srgb-linear 1 0.5 0.2 / 0.6)");
    testColorFunction("Linear-light sRGB with alpha, mixed number and percent 2", "color(srgb-linear 1.00 50% 0.2 / 60%)", "color(srgb-linear 1 0.5 0.2 / 0.6)");
    testColorFunction("Linear-light sRGB with alpha, all none", "color(srgb-linear none none none / none)", "color(srgb-linear none none none / none)");
    testColorFunction("Linear-light sRGB with alpha, number and none", "color(srgb-linear 1.00 none 0.2 / none)", "color(srgb-linear 1 none 0.2 / none)");
    testColorFunction("Linear-light sRGB with alpha, percent and none", "color(srgb-linear 100% none 20% / 30%)", "color(srgb-linear 1 none 0.2 / 0.3)");
    testColorFunction("Linear-light sRGB with alpha, number, percent and none", "color(srgb-linear 100% none 0.2 / 23.7%)", "color(srgb-linear 1 none 0.2 / 0.237)");

    // Opaque Display P3 in color()
    testColorFunction("Display P3 all numbers", "color(display-p3 1.00 0.50 0.200)", "color(display-p3 1 0.5 0.2)");
    testColorFunction("Display P3 all percent", "color(display-p3 100% 50% 20%)", "color(display-p3 1 0.5 0.2)");
    testColorFunction("Display P3 mixed number and percent", "color(display-p3 100% 0.5 20%)", "color(display-p3 1 0.5 0.2)");
    testColorFunction("Display P3 mixed number and percent 2", "color(display-p3 1.00 50% 0.2)", "color(display-p3 1 0.5 0.2)");
    testColorFunction("Display P3 all none", "color(display-p3 none none none)", "color(display-p3 none none none)");
    testColorFunction("Display P3 number and none", "color(display-p3 1.00 none 0.2)", "color(display-p3 1 none 0.2)");
    testColorFunction("Display P3 percent and none", "color(display-p3 100% none 20%)", "color(display-p3 1 none 0.2)");
    testColorFunction("Display P3 number, percent and none", "color(display-p3 100% none 0.2)", "color(display-p3 1 none 0.2)");

    // non-unity alpha, Display P3 in  color()
    testColorFunction("Display P3 with alpha, all numbers", "color(display-p3 1.00 0.50 0.200 / 0.6)", "color(display-p3 1 0.5 0.2 / 0.6)");
    testColorFunction("Display P3 with alpha, all percent", "color(display-p3 100% 50% 20% / 60%)", "color(display-p3 1 0.5 0.2 / 0.6)");
    testColorFunction("Display P3 with alpha, mixed number and percent", "color(display-p3 100% 0.5 20% / 0.6)", "color(display-p3 1 0.5 0.2 / 0.6)");
    testColorFunction("Display P3 with alpha, mixed number and percent 2", "color(display-p3 1.00 50% 0.2 / 60%)", "color(display-p3 1 0.5 0.2 / 0.6)");
    testColorFunction("Display P3 with alpha, all none", "color(display-p3 none none none / none)", "color(display-p3 none none none / none)");
    testColorFunction("Display P3 with alpha, number and none", "color(display-p3 1.00 none 0.2 / none)", "color(display-p3 1 none 0.2 / none)");
    testColorFunction("Display P3 with alpha, percent and none", "color(display-p3 100% none 20% / 30%)", "color(display-p3 1 none 0.2 / 0.3)");
    testColorFunction("Display P3 with alpha, number, percent and none", "color(display-p3 100% none 0.2 / 23.7%)", "color(display-p3 1 none 0.2 / 0.237)");

    // Opaque A98 RGB in color()
    testColorFunction("A98 RGB all numbers", "color(a98-rgb 1.00 0.50 0.200)", "color(a98-rgb 1 0.5 0.2)");
    testColorFunction("A98 RGB all percent", "color(a98-rgb 100% 50% 20%)", "color(a98-rgb 1 0.5 0.2)");
    testColorFunction("A98 RGB mixed number and percent", "color(a98-rgb 100% 0.5 20%)", "color(a98-rgb 1 0.5 0.2)");
    testColorFunction("A98 RGB mixed number and percent 2", "color(a98-rgb 1.00 50% 0.2)", "color(a98-rgb 1 0.5 0.2)");
    testColorFunction("A98 RGB all none", "color(a98-rgb none none none)", "color(a98-rgb none none none)");
    testColorFunction("A98 RGB number and none", "color(a98-rgb 1.00 none 0.2)", "color(a98-rgb 1 none 0.2)");
    testColorFunction("A98 RGB percent and none", "color(a98-rgb 100% none 20%)", "color(a98-rgb 1 none 0.2)");
    testColorFunction("A98 RGB number, percent and none", "color(a98-rgb 100% none 0.2)", "color(a98-rgb 1 none 0.2)");

    // non-unity alpha, A98 RGB in  color()
    testColorFunction("A98 RGB with alpha, all numbers", "color(a98-rgb 1.00 0.50 0.200 / 0.6)", "color(a98-rgb 1 0.5 0.2 / 0.6)");
    testColorFunction("A98 RGB with alpha, all percent", "color(a98-rgb 100% 50% 20% / 60%)", "color(a98-rgb 1 0.5 0.2 / 0.6)");
    testColorFunction("A98 RGB with alpha, mixed number and percent", "color(a98-rgb 100% 0.5 20% / 0.6)", "color(a98-rgb 1 0.5 0.2 / 0.6)");
    testColorFunction("A98 RGB with alpha, mixed number and percent 2", "color(a98-rgb 1.00 50% 0.2 / 60%)", "color(a98-rgb 1 0.5 0.2 / 0.6)");
    testColorFunction("A98 RGB with alpha, all none", "color(a98-rgb none none none / none)", "color(a98-rgb none none none / none)");
    testColorFunction("A98 RGB with alpha, number and none", "color(a98-rgb 1.00 none 0.2 / none)", "color(a98-rgb 1 none 0.2 / none)");
    testColorFunction("A98 RGB with alpha, percent and none", "color(a98-rgb 100% none 20% / 30%)", "color(a98-rgb 1 none 0.2 / 0.3)");
    testColorFunction("A98 RGB with alpha, number, percent and none", "color(a98-rgb 100% none 0.2 / 23.7%)", "color(a98-rgb 1 none 0.2 / 0.237)");

    // Opaque ProPhoto RGB in color()
    testColorFunction("ProPhoto RGB all numbers", "color(prophoto-rgb 1.00 0.50 0.200)", "color(prophoto-rgb 1 0.5 0.2)");
    testColorFunction("ProPhoto RGB all percent", "color(prophoto-rgb 100% 50% 20%)", "color(prophoto-rgb 1 0.5 0.2)");
    testColorFunction("ProPhoto RGB mixed number and percent", "color(prophoto-rgb 100% 0.5 20%)", "color(prophoto-rgb 1 0.5 0.2)");
    testColorFunction("ProPhoto RGB mixed number and percent 2", "color(prophoto-rgb 1.00 50% 0.2)", "color(prophoto-rgb 1 0.5 0.2)");
    testColorFunction("ProPhoto RGB all none", "color(prophoto-rgb none none none)", "color(prophoto-rgb none none none)");
    testColorFunction("ProPhoto RGB number and none", "color(prophoto-rgb 1.00 none 0.2)", "color(prophoto-rgb 1 none 0.2)");
    testColorFunction("ProPhoto RGB percent and none", "color(prophoto-rgb 100% none 20%)", "color(prophoto-rgb 1 none 0.2)");
    testColorFunction("ProPhoto RGB number, percent and none", "color(prophoto-rgb 100% none 0.2)", "color(prophoto-rgb 1 none 0.2)");

    // non-unity alpha, ProPhoto RGB in  color()
    testColorFunction("ProPhoto RGB with alpha, all numbers", "color(prophoto-rgb 1.00 0.50 0.200 / 0.6)", "color(prophoto-rgb 1 0.5 0.2 / 0.6)");
    testColorFunction("ProPhoto RGB with alpha, all percent", "color(prophoto-rgb 100% 50% 20% / 60%)", "color(prophoto-rgb 1 0.5 0.2 / 0.6)");
    testColorFunction("ProPhoto RGB with alpha, mixed number and percent", "color(prophoto-rgb 100% 0.5 20% / 0.6)", "color(prophoto-rgb 1 0.5 0.2 / 0.6)");
    testColorFunction("ProPhoto RGB with alpha, mixed number and percent 2", "color(prophoto-rgb 1.00 50% 0.2 / 60%)", "color(prophoto-rgb 1 0.5 0.2 / 0.6)");
    testColorFunction("ProPhoto RGB with alpha, all none", "color(prophoto-rgb none none none / none)", "color(prophoto-rgb none none none / none)");
    testColorFunction("ProPhoto RGB with alpha, number and none", "color(prophoto-rgb 1.00 none 0.2 / none)", "color(prophoto-rgb 1 none 0.2 / none)");
    testColorFunction("ProPhoto RGB with alpha, percent and none", "color(prophoto-rgb 100% none 20% / 30%)", "color(prophoto-rgb 1 none 0.2 / 0.3)");
    testColorFunction("ProPhoto RGB with alpha, number, percent and none", "color(prophoto-rgb 100% none 0.2 / 23.7%)", "color(prophoto-rgb 1 none 0.2 / 0.237)");

    // Opaque Rec BT.2020 in  color()
    testColorFunction("Rec BT.2020 all numbers", "color(rec2020 1.00 0.50 0.200)", "color(rec2020 1 0.5 0.2)");
    testColorFunction("Rec BT.2020 all percent", "color(rec2020 100% 50% 20%)", "color(rec2020 1 0.5 0.2)");
    testColorFunction("Rec BT.2020 mixed number and percent", "color(rec2020 100% 0.5 20%)", "color(rec2020 1 0.5 0.2)");
    testColorFunction("Rec BT.2020 mixed number and percent 2", "color(rec2020 1.00 50% 0.2)", "color(rec2020 1 0.5 0.2)");
    testColorFunction("Rec BT.2020 all none", "color(rec2020 none none none)", "color(rec2020 none none none)");
    testColorFunction("Rec BT.2020 number and none", "color(rec2020 1.00 none 0.2)", "color(rec2020 1 none 0.2)");
    testColorFunction("Rec BT.2020 percent and none", "color(rec2020 100% none 20%)", "color(rec2020 1 none 0.2)");
    testColorFunction("Rec BT.2020 number, percent and none", "color(rec2020 100% none 0.2)", "color(rec2020 1 none 0.2)");

    // non-unity alpha, Rec BT.2020 in  color()
    testColorFunction("Rec BT.2020 with alpha, all numbers", "color(rec2020 1.00 0.50 0.200 / 0.6)", "color(rec2020 1 0.5 0.2 / 0.6)");
    testColorFunction("Rec BT.2020 with alpha, all percent", "color(rec2020 100% 50% 20% / 60%)", "color(rec2020 1 0.5 0.2 / 0.6)");
    testColorFunction("Rec BT.2020 with alpha, mixed number and percent", "color(rec2020 100% 0.5 20% / 0.6)", "color(rec2020 1 0.5 0.2 / 0.6)");
    testColorFunction("Rec BT.2020 with alpha, mixed number and percent 2", "color(rec2020 1.00 50% 0.2 / 60%)", "color(rec2020 1 0.5 0.2 / 0.6)");
    testColorFunction("Rec BT.2020 with alpha, all none", "color(rec2020 none none none / none)", "color(rec2020 none none none / none)");
    testColorFunction("Rec BT.2020 with alpha, number and none", "color(rec2020 1.00 none 0.2 / none)", "color(rec2020 1 none 0.2 / none)");
    testColorFunction("Rec BT.2020 with alpha, percent and none", "color(rec2020 100% none 20% / 30%)", "color(rec2020 1 none 0.2 / 0.3)");
    testColorFunction("Rec BT.2020 with alpha, number, percent and none", "color(rec2020 100% none 0.2 / 23.7%)", "color(rec2020 1 none 0.2 / 0.237)");

    // Opaque CIE XYZ D50 in color()
    testColorFunction("CIE XYZ D50 all numbers", "color(xyz-d50 1.00 0.50 0.200)", "color(xyz-d50 1 0.5 0.2)");
    testColorFunction("CIE XYZ D50 all percent", "color(xyz-d50 100% 50% 20%)", "color(xyz-d50 1 0.5 0.2)");
    testColorFunction("CIE XYZ D50 mixed number and percent", "color(xyz-d50 100% 0.5 20%)", "color(xyz-d50 1 0.5 0.2)");
    testColorFunction("CIE XYZ D50 mixed number and percent 2", "color(xyz-d50 1.00 50% 0.2)", "color(xyz-d50 1 0.5 0.2)");
    testColorFunction("CIE XYZ D50 all none", "color(xyz-d50 none none none)", "color(xyz-d50 none none none)");
    testColorFunction("CIE XYZ D50 number and none", "color(xyz-d50 1.00 none 0.2)", "color(xyz-d50 1 none 0.2)");
    testColorFunction("CIE XYZ D50 percent and none", "color(xyz-d50 100% none 20%)", "color(xyz-d50 1 none 0.2)");
    testColorFunction("CIE XYZ D50 number, percent and none", "color(xyz-d50 100% none 0.2)", "color(xyz-d50 1 none 0.2)");

    // non-unity alpha, CIE XYZ D50 in  color()
    testColorFunction("CIE XYZ D50 with alpha, all numbers", "color(xyz-d50 1.00 0.50 0.200 / 0.6)", "color(xyz-d50 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ D50 with alpha, all percent", "color(xyz-d50 100% 50% 20% / 60%)", "color(xyz-d50 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ D50 with alpha, mixed number and percent", "color(xyz-d50 100% 0.5 20% / 0.6)", "color(xyz-d50 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ D50 with alpha, mixed number and percent 2", "color(xyz-d50 1.00 50% 0.2 / 60%)", "color(xyz-d50 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ D50 with alpha, all none", "color(xyz-d50 none none none / none)", "color(xyz-d50 none none none / none)");
    testColorFunction("CIE XYZ D50 with alpha, number and none", "color(xyz-d50 1.00 none 0.2 / none)", "color(xyz-d50 1 none 0.2 / none)");
    testColorFunction("CIE XYZ D50 with alpha, percent and none", "color(xyz-d50 100% none 20% / 30%)", "color(xyz-d50 1 none 0.2 / 0.3)");
    testColorFunction("CIE XYZ D50 with alpha, number, percent and none", "color(xyz-d50 100% none 0.2 / 23.7%)", "color(xyz-d50 1 none 0.2 / 0.237)");

    // Opaque CIE XYZ D65 in color()
    testColorFunction("CIE XYZ D65 all numbers", "color(xyz-d65 1.00 0.50 0.200)", "color(xyz-d65 1 0.5 0.2)");
    testColorFunction("CIE XYZ D65 all percent", "color(xyz-d65 100% 50% 20%)", "color(xyz-d65 1 0.5 0.2)");
    testColorFunction("CIE XYZ D65 mixed number and percent", "color(xyz-d65 100% 0.5 20%)", "color(xyz-d65 1 0.5 0.2)");
    testColorFunction("CIE XYZ D65 mixed number and percent 2", "color(xyz-d65 1.00 50% 0.2)", "color(xyz-d65 1 0.5 0.2)");
    testColorFunction("CIE XYZ D65 all none", "color(xyz-d65 none none none)", "color(xyz-d65 none none none)");
    testColorFunction("CIE XYZ D65 number and none", "color(xyz-d65 1.00 none 0.2)", "color(xyz-d65 1 none 0.2)");
    testColorFunction("CIE XYZ D65 percent and none", "color(xyz-d65 100% none 20%)", "color(xyz-d65 1 none 0.2)");
    testColorFunction("CIE XYZ D65 number, percent and none", "color(xyz-d65 100% none 0.2)", "color(xyz-d65 1 none 0.2)");

    // non-unity alpha, CIE XYZ D65 in  color()
    testColorFunction("CIE XYZ D65 with alpha, all numbers", "color(xyz-d65 1.00 0.50 0.200 / 0.6)", "color(xyz-d65 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ D65 with alpha, all percent", "color(xyz-d65 100% 50% 20% / 60%)", "color(xyz-d65 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ D65 with alpha, mixed number and percent", "color(xyz-d65 100% 0.5 20% / 0.6)", "color(xyz-d65 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ D65 with alpha, mixed number and percent 2", "color(xyz-d65 1.00 50% 0.2 / 60%)", "color(xyz-d65 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ D65 with alpha, all none", "color(xyz-d65 none none none / none)", "color(xyz-d65 none none none / none)");
    testColorFunction("CIE XYZ D65 with alpha, number and none", "color(xyz-d65 1.00 none 0.2 / none)", "color(xyz-d65 1 none 0.2 / none)");
    testColorFunction("CIE XYZ D65 with alpha, percent and none", "color(xyz-d65 100% none 20% / 30%)", "color(xyz-d65 1 none 0.2 / 0.3)");
    testColorFunction("CIE XYZ D65 with alpha, number, percent and none", "color(xyz-d65 100% none 0.2 / 23.7%)", "color(xyz-d65 1 none 0.2 / 0.237)");

    // Opaque CIE XYZ (implicit D65) in color()
    testColorFunction("CIE XYZ (implicit D65) all numbers", "color(xyz 1.00 0.50 0.200)", "color(xyz-d65 1 0.5 0.2)");
    testColorFunction("CIE XYZ (implicit D65) all percent", "color(xyz 100% 50% 20%)", "color(xyz-d65 1 0.5 0.2)");
    testColorFunction("CIE XYZ (implicit D65) mixed number and percent", "color(xyz 100% 0.5 20%)", "color(xyz-d65 1 0.5 0.2)");
    testColorFunction("CIE XYZ (implicit D65) mixed number and percent 2", "color(xyz 1.00 50% 0.2)", "color(xyz-d65 1 0.5 0.2)");
    testColorFunction("CIE XYZ (implicit D65) all none", "color(xyz none none none)", "color(xyz-d65 none none none)");
    testColorFunction("CIE XYZ (implicit D65) number and none", "color(xyz 1.00 none 0.2)", "color(xyz-d65 1 none 0.2)");
    testColorFunction("CIE XYZ (implicit D65) percent and none", "color(xyz 100% none 20%)", "color(xyz-d65 1 none 0.2)");
    testColorFunction("CIE XYZ (implicit D65) number, percent and none", "color(xyz 100% none 0.2)", "color(xyz-d65 1 none 0.2)");

    // non-unity alpha, CIE XYZ (implicit D65) in  color()
    testColorFunction("CIE XYZ (implicit D65) with alpha, all numbers", "color(xyz 1.00 0.50 0.200 / 0.6)", "color(xyz-d65 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ (implicit D65) with alpha, all percent", "color(xyz 100% 50% 20% / 60%)", "color(xyz-d65 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ (implicit D65) with alpha, mixed number and percent", "color(xyz 100% 0.5 20% / 0.6)", "color(xyz-d65 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ (implicit D65) with alpha, mixed number and percent 2", "color(xyz 1.00 50% 0.2 / 60%)", "color(xyz-d65 1 0.5 0.2 / 0.6)");
    testColorFunction("CIE XYZ (implicit D65) with alpha, all none", "color(xyz none none none / none)", "color(xyz-d65 none none none / none)");
    testColorFunction("CIE XYZ (implicit D65) with alpha, number and none", "color(xyz 1.00 none 0.2 / none)", "color(xyz-d65 1 none 0.2 / none)");
    testColorFunction("CIE XYZ (implicit D65) with alpha, percent and none", "color(xyz 100% none 20% / 30%)", "color(xyz-d65 1 none 0.2 / 0.3)");
    testColorFunction("CIE XYZ (implicit D65) with alpha, number, percent and none", "color(xyz 100% none 0.2 / 23.7%)", "color(xyz-d65 1 none 0.2 / 0.237)");

</script>
